Change pong-server file architecture to properly separate the different programs
