+++ pong_server/.gitignore
@@ -1,0 +1,3 @@
+src/game_creator/shared_code
+src/game_server/shared_code
+src/redirection_server_deprecated/shared_code
