+++ pong_server/dev_deprecated_redirection_server.sh
@@ -14,7 +14,11 @@ while true; do
     kill -9 $(pgrep '[P]ython')
     sleep 1
     clear
-    python -m unittest discover test
     `sed 's/^/export /' ../.env`
-    find . -name '*.py' | entr -z -d python3 -m src.redirection_server_deprecated
+    python -m unittest discover test
+    (
+        cd src/redirection_server_deprecated &&
+        find . -name '*.py' |
+        entr -z -d python3 main.py
+    )
 done
