+++ pong_server/dev_game_creator.sh
@@ -2,7 +2,7 @@
 
 export PONG_GAME_SERVERS_MIN_PORT=42200
 export PONG_GAME_SERVERS_MAX_PORT=42210
-export PONG_GAME_APP_PATH=~/git/transcendence/pong_server/
+export GAME_SERVER_PATH=~/git/transcendence/pong_server/src/game_server/
 
 
 (cd src/game_creator/ && python3 manage.py test) || exit $?
