+++ pong_server/src/game_creator/api/urls.py
@@ -1,5 +1,6 @@
-from api.views import CreateGameView
 from django.urls import path
+
+from api.views import CreateGameView
 
 urlpatterns = [
     path('', CreateGameView.as_view(), name='create_game'),
